chore: make liquidity page client side only
