--- apps/trading/pages/liquidity/index.ts
@@ -1,1 +1,0 @@
-export * from './[marketId].page';
